Refactor Constants, Remove Uneccessary Javadoc Comments and rename updateInterface.java to VersionManager.java
